# docs/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact: {}
  description: 荆楚花牌(花好月圆)游戏服务端 API
  title: 荆楚花牌 API
  version: "1.0"
paths:
...
  /games/{id}:
    get:
      description: 获取指定游戏的完整状态
      parameters:
      - description: 游戏 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 游戏状态
          schema:
            type: object
        "404":
          description: '{ \"error\": \"game not found\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 查询游戏状态
      tags:
      - 游戏
  /games/{id}/cut:
    post:
      consumes:
      - application/json
      description: 歇家或闲二选择切牌位置（37-110）
      parameters:
      - description: 游戏 ID
        in: path
        name: id
        required: true
        type: string
      - description: 切牌位置
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{ \"status\": \"ok\" }'
          schema:
            type: object
        "400":
          description: '{ \"error\": \"string\" }'
          schema:
            type: object
        "404":
          description: '{ \"error\": \"game not found\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 切牌（腰牌）
      tags:
      - 游戏
  /games/{id}/dang-jing:
    post:
      consumes:
      - application/json
      description: 选择当经牌（三、五、七之一）
      parameters:
      - description: 游戏 ID
        in: path
        name: id
        required: true
        type: string
      - description: 当经选择
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{ \"status\": \"ok\" }'
          schema:
            type: object
        "400":
          description: '{ \"error\": \"string\" }'
          schema:
            type: object
        "404":
          description: '{ \"error\": \"game not found\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 选择当经
      tags:
      - 游戏
  /games/{id}/discard:
    post:
      consumes:
      - application/json
      description: 打出一张手牌
      parameters:
      - description: 游戏 ID
        in: path
        name: id
        required: true
        type: string
      - description: 出牌信息
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{ \"status\": \"ok\" }'
          schema:
            type: object
        "400":
          description: '{ \"error\": \"string\" }'
          schema:
            type: object
        "404":
          description: '{ \"error\": \"game not found\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 出牌
      tags:
      - 游戏
  /games/{id}/draw:
    post:
      description: 从公牌顶部起一张牌
      parameters:
      - description: 游戏 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{ \"status\": \"ok\" }'
          schema:
            type: object
        "400":
          description: '{ \"error\": \"string\" }'
          schema:
            type: object
        "404":
          description: '{ \"error\": \"game not found\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 起牌
      tags:
      - 游戏
  /games/{id}/ganta:
    post:
      description: 将手中同名牌加入已有的统组合中
      parameters:
      - description: 游戏 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{ \"status\": \"ok\" }'
          schema:
            type: object
        "400":
          description: '{ \"error\": \"string\" }'
          schema:
            type: object
        "404":
          description: '{ \"error\": \"game not found\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 赶塔
      tags:
      - 游戏
  /games/{id}/pair:
    post:
      consumes:
      - application/json
      description: 对别人打出的牌进行碰牌操作
      parameters:
      - description: 游戏 ID
        in: path
        name: id
        required: true
        type: string
      - description: 碰牌信息
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{ \"status\": \"ok\" }'
          schema:
            type: object
        "400":
          description: '{ \"error\": \"string\" }'
          schema:
            type: object
        "404":
          description: '{ \"error\": \"game not found\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 碰牌（对/招/泛）
      tags:
      - 游戏
  /games/{id}/pass:
    post:
      description: 放弃当前操作（碰牌、和牌等）
      parameters:
      - description: 游戏 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{ \"status\": \"ok\" }'
          schema:
            type: object
        "400":
          description: '{ \"error\": \"string\" }'
          schema:
            type: object
        "404":
          description: '{ \"error\": \"game not found\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 过
      tags:
      - 游戏
  /games/{id}/tong:
    post:
      consumes:
      - application/json
      description: 请统阶段，玩家选择统牌或跳过
      parameters:
      - description: 游戏 ID
        in: path
        name: id
        required: true
        type: string
      - description: 统牌信息
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{ \"status\": \"ok\" }'
          schema:
            type: object
        "400":
          description: '{ \"error\": \"string\" }'
          schema:
            type: object
        "404":
          description: '{ \"error\": \"game not found\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 统牌抉择
      tags:
      - 游戏
  /games/{id}/win:
    post:
      description: 声明和牌，系统验证手牌是否满足和牌条件
      parameters:
      - description: 游戏 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{ \"status\": \"ok\" }'
          schema:
            type: object
        "400":
          description: '{ \"error\": \"string\" }'
          schema:
            type: object
        "404":
          description: '{ \"error\": \"game not found\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 和牌（胡牌）
      tags:
      - 游戏
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"